refactor(012): split quiz-table.ts into focused units (US3)
Decompose the 824-line quiz-table.ts (now 399) into:
- quiz-table-columns.ts: answer/detail column show/hide + removeColgroup
- quiz-input-factory.ts: createQuestionInput
- quiz-answer-persistence.ts: handleAnswerInput/saveAnswer/applyValidationStyling
- quiz-instructor-overlay.ts: show/hide student answers

quiz-table.ts retains only lifecycle/orchestration. Re-point instructor-
answer-reveal and the overlay integration test to the new overlay module.
Existing quiz-table integration suite is the characterization net (T023).
